% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes.R
\name{aes_decrypt}
\alias{aes_decrypt}
\title{AES Decryption Wrapper Function}
\usage{
aes_decrypt(x, iv, key, output_format = "string", input_format = "hex")
}
\arguments{
\item{x}{The input to be decrypted, either a character string or a raw vector.}

\item{iv}{A raw vector representing the initialization vector (IV).}

\item{key}{The decryption key as a character string or raw vector. It will be hashed to 256 bits if provided as a string.}

\item{output_format}{The format of the output. Options are "string" (default), "hex", or "none" for raw bytes.}

\item{input_format}{The format of the encrypted input. Options are "hex" (default) or "base64".}
}
\value{
The decrypted data in the specified format.
}
\description{
Decrypts the input using AES decryption in CBC mode with PKCS7 padding, with default input format as hexadecimal and output format as a string.
}
\examples{
encrypted_hex <- aes_encrypt("Encrypt this message!", charToRaw("thisisaniv123456"), "mysecretpassword")
aes_decrypt(encrypted_hex, charToRaw("thisisaniv123456"), "mysecretpassword")

}
