% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes.R
\name{aes_encrypt}
\alias{aes_encrypt}
\title{Prepare data for AES encryption}
\usage{
aes_encrypt(
  x,
  key,
  iv = c(6, 224, 71, 170, 241, 204, 115, 21, 30, 8, 46, 223, 106, 207, 55, 42),
  output_format = "hex"
)
}
\arguments{
\item{x}{The input data to encrypt. This can be a character string or a raw vector.}

\item{key}{The encryption key. This can be a character string (hashed into a 256-bit key) or a raw vector.}

\item{iv}{An optional vector of unsigned bytes of size 16 to be used as the initialization vector. Defaults to a predefined IV.}

\item{output_format}{The desired format for the encrypted output: "hex", "base64" or "none". Defaults to "hex".}
}
\value{
The encrypted data in the specified output format.
}
\description{
This function does the necessary type-checking and conversions of parameters
and then passes them to "encrypt_aes_cbc" for AES encryption.
It also transforms the result to the specified/default output format.
}
