% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jws.R
\name{b64}
\alias{b64}
\title{Base64 URL encode}
\usage{
b64(input_string)
}
\arguments{
\item{input_string}{A string to be encoded.}
}
\value{
A Base64 URL encoded string without padding.
}
\description{
This function encodes a given string into Base64 URL format and removes any
trailing '=' padding. It is useful for encoding data in a URL-safe way
without padding characters. Used internally for JWS serialization.
}
\examples{
\dontrun{
b64("example string")
}

}
\keyword{internal}
