% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encodings.R
\name{biginteger_to_bytes}
\alias{biginteger_to_bytes}
\title{Converts a big integer to bytes}
\usage{
biginteger_to_bytes(bn, len = NULL)
}
\arguments{
\item{bn}{A big integer value (can be numeric or character)}

\item{len}{Optional length of byte array (if not provided, the minimum length is used)}
}
\value{
A raw byte vector representing the big integer
}
\description{
Converts a big integer to bytes
}
