% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encodings.R
\name{biginteger_to_hex}
\alias{biginteger_to_hex}
\title{Convert a big integer to a hexadecimal string}
\usage{
biginteger_to_hex(bn)
}
\arguments{
\item{bn}{A big integer value (as an R numeric or character)}
}
\value{
A string representing the hexadecimal value of the big integer
}
\description{
Convert a big integer to a hexadecimal string
}
