% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkcs7.R
\name{byte_check}
\alias{byte_check}
\title{Check if a value is a valid byte}
\usage{
byte_check(b)
}
\arguments{
\item{b}{A number to check.}
}
\value{
TRUE if the value is a valid byte, otherwise FALSE.
}
\description{
Determines if a given value is a valid byte (an integer between 0 and 255).
}
