% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encodings.R
\name{byte_to_int}
\alias{byte_to_int}
\title{Convert the first byte of a raw byte vector to an integer}
\usage{
byte_to_int(the_bytes)
}
\arguments{
\item{the_bytes}{A raw byte vector}
}
\value{
An integer representation of the first byte
}
\description{
Convert the first byte of a raw byte vector to an integer
}
