% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkcs7.R
\name{bytes_check}
\alias{bytes_check}
\title{Check if a vector contains valid bytes}
\usage{
bytes_check(bytes)
}
\arguments{
\item{bytes}{A numeric vector.}
}
\value{
TRUE if all elements are valid bytes, otherwise FALSE.
}
\description{
Determines if all elements in a vector are valid bytes.
}
