% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkcs7.R
\name{compare_bytes}
\alias{compare_bytes}
\title{Compare two byte arrays}
\usage{
compare_bytes(ba1, ba2)
}
\arguments{
\item{ba1}{A numeric vector (byte array).}

\item{ba2}{A numeric vector (byte array).}
}
\value{
TRUE if the byte arrays are identical, otherwise FALSE.
}
\description{
Compares two byte arrays for equality. Returns FALSE if they are of different lengths or if either contains invalid bytes.
}
