% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secp256k1_wrapper.R
\name{compute_recovery_byte}
\alias{compute_recovery_byte}
\title{Compute Recovery Byte}
\usage{
compute_recovery_byte(kp, r, s)
}
\arguments{
\item{kp}{An object representing the elliptic curve point.}

\item{r}{A numeric representing the R parameter of the ECDSA signature.}

\item{s}{A numeric representing the S parameter of the ECDSA signature.}
}
\value{
An integer representing the recovery byte.
}
\description{
Compute a recovery byte for a compressed ECDSA signature given R and S parameters.
Returns the value as a byte integer.
}
