% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sepc256k1_curve.R
\name{containsPoint}
\alias{containsPoint}
\title{Check if a Point is on the Elliptic Curve}
\usage{
containsPoint(curve, x, y)
}
\arguments{
\item{curve}{An object of class ECPARAM representing the elliptic curve parameters.}

\item{x}{A big integer or hexadecimal string representing the x-coordinate of the point.}

\item{y}{A big integer or hexadecimal string representing the y-coordinate of the point.}
}
\value{
A logical value: TRUE if the point is on the curve, FALSE otherwise.
}
\description{
This function checks whether a point (x, y) lies on the elliptic curve defined by the given parameters.
}
\examples{
p <- "0xFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFEFFFFFC2F"
b <- "0x0000000000000000000000000000000000000000000000000000000000000007"
a <- "0x0000000000000000000000000000000000000000000000000000000000000000"
x <- "0x79BE667EF9DCBBAC55A06295CE870B07029BFCDB2DCE28D959F2815B16F81798"
y <- "0x483ada7726a3c4655da4fbfc0e1108a8fd17b448a68554199c47d08ffb10d4b8"
curve256 <- ecparam(p, a, b)
is_on_curve <- containsPoint(curve256, x, y)
print(is_on_curve)

}
