% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_valid_secp256k1PK.R
\name{create_valid_secp256k1_private_key}
\alias{create_valid_secp256k1_private_key}
\title{create_valid_secp256k1_private_key}
\usage{
create_valid_secp256k1_private_key()
}
\value{
A hexadecimal string representing a valid private key for the secp256k1 curve.
}
\description{
create_valid_secp256k1_private_key
}
\examples{
private_key <- create_valid_secp256k1_private_key()
print(private_key)
private_bn <- gmp::as.bigz(paste0("0x", private_key))
is_valid <- valid_private_key(private_bn)
print(is_valid)  # This should now print TRUE
}
