% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asn1.R
\name{decode_asn1_integer}
\alias{decode_asn1_integer}
\title{Decode ASN.1 Integer}
\usage{
decode_asn1_integer(asn1)
}
\arguments{
\item{asn1}{A string containing the ASN.1 encoded data.}
}
\value{
A list with two elements: the decoded integer and the remaining ASN.1 string.
}
\description{
Decodes an integer from the top of an ASN.1 encoded string. The function assumes the ASN.1 integer tag "02" is present.
}
