% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asn1.R
\name{decode_asn1_length}
\alias{decode_asn1_length}
\title{Decode ASN.1 Length}
\usage{
decode_asn1_length(asn1)
}
\arguments{
\item{asn1}{A string containing the ASN.1 encoded data.}
}
\value{
A list with two elements: the decoded length and the remaining ASN.1 string.
}
\description{
Decodes the length from an ASN.1 encoded string, throwing an error for unsupported lengths.
}
