% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sepc256k1_curve.R
\name{ecparam}
\alias{ecparam}
\title{Create an Elliptic Curve Parameter Object}
\usage{
ecparam(p, a, b)
}
\arguments{
\item{p}{A big integer or hexadecimal string representing the prime modulus of the curve.}

\item{a}{A big integer or hexadecimal string representing the coefficient a of the curve equation.}

\item{b}{A big integer or hexadecimal string representing the coefficient b of the curve equation.}
}
\value{
An object of class ECPARAM containing the curve parameters.
}
\description{
This function creates an elliptic curve parameter object for use with elliptic curve operations.
}
\examples{
p <- "0xFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFEFFFFFC2F"
b <- "0x0000000000000000000000000000000000000000000000000000000000000007"
a <- "0x0000000000000000000000000000000000000000000000000000000000000000"
curve256 <- ecparam(p, a, b)
print(curve256)

}
