% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asn1.R
\name{encode_asn1_length_hex}
\alias{encode_asn1_length_hex}
\title{Encode ASN.1 Length to Hexadecimal}
\usage{
encode_asn1_length_hex(len)
}
\arguments{
\item{len}{An integer representing the length to encode.}
}
\value{
A string containing the hexadecimal representation of the length.
}
\description{
Converts a length into its hexadecimal string representation for ASN.1 encoding.
}
