% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asn1.R
\name{encode_asn1_unsigned_integer}
\alias{encode_asn1_unsigned_integer}
\title{Encode ASN.1 Unsigned Integer (Byte Array)}
\usage{
encode_asn1_unsigned_integer(ba)
}
\arguments{
\item{ba}{A byte array representing an unsigned integer.}
}
\value{
A byte array with the ASN.1 encoded unsigned integer.
}
\description{
Encodes a byte array as an ASN.1 unsigned integer, including length and ASN.1 integer tag.
}
