% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asn1.R
\name{encode_asn1_unsigned_integer_hex}
\alias{encode_asn1_unsigned_integer_hex}
\title{Encode ASN.1 Unsigned Integer (Hexadecimal)}
\usage{
encode_asn1_unsigned_integer_hex(n)
}
\arguments{
\item{n}{A hexadecimal string representing an unsigned integer.}
}
\value{
A string containing the ASN.1 encoded unsigned integer.
}
\description{
Encodes a hexadecimal string as an ASN.1 unsigned integer, including length and ASN.1 integer tag.
}
