% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes.R, R/scrypt.R
\name{encrypt}
\alias{encrypt}
\title{Prepare data for AES encryption}
\usage{
encrypt(raw, salt = random_bytes(16), n = 32768, r = 8, p = 1, dk_len = 32)

encrypt(raw, salt = random_bytes(16), n = 32768, r = 8, p = 1, dk_len = 32)
}
\arguments{
\item{raw}{A raw vector containing the message to encrypt.}

\item{salt}{A raw vector containing the salt to use (default random 16 bytes).}

\item{n}{Integer. CPU/memory cost factor (default 32768).}

\item{r}{Integer. Block size factor (default 8).}

\item{p}{Integer. Parallelization factor (default 1).}

\item{dk_len}{Integer. Length of the derived key (default 32).}

\item{x}{The input data to encrypt. This can be a character string or a raw vector.}

\item{key}{The encryption key. This can be a character string (hashed into a 256-bit key) or a raw vector.}

\item{iv}{An optional vector of unsigned bytes of size 16 to be used as the initialization vector. Defaults to a predefined IV.}

\item{output_format}{The desired format for the encrypted output: "hex", "base64" or "none". Defaults to "hex".}
}
\value{
The encrypted data in the specified output format.

A raw vector containing the encrypted message.
}
\description{
This function does the necessary type-checking and conversions of parameters
and then passes them to "encrypt_aes_cbc" for AES encryption.
It also transforms the result to the specified/default output format.

Encrypts a message (raw bytes) using a specified salt and scrypt parameters.
Returns the encrypted message in bytes directly. The encrypted output can be
verified using the same salt and scrypt parameters.
}
