% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes.R
\name{encrypt_aes_cbc}
\alias{encrypt_aes_cbc}
\title{AES Encryption with CBC Mode}
\usage{
encrypt_aes_cbc(iv, key, data)
}
\arguments{
\item{iv}{A raw vector representing the initialization vector (IV).}

\item{key}{A raw vector representing the key for AES encryption. It should be exactly 16, 24, or 32 bytes.}

\item{data}{A raw vector representing the data to be encrypted.}
}
\value{
A raw vector representing the encrypted data.
}
\description{
This function encrypts data using AES encryption in CBC mode with PKCS7 padding (compatible with PKCS5).
}
