% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asn1.R
\name{format_asn1_unsigned_integer}
\alias{format_asn1_unsigned_integer}
\title{Format ASN.1 Unsigned Integer (Byte Array)}
\usage{
format_asn1_unsigned_integer(ba)
}
\arguments{
\item{ba}{A byte array representing an unsigned integer.}
}
\value{
A byte array with formatted unsigned integer.
}
\description{
Formats a byte array representing an unsigned integer by removing leading zeros and padding if necessary.
}
