% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asn1.R
\name{format_asn1_unsigned_integer_hex}
\alias{format_asn1_unsigned_integer_hex}
\title{Format ASN.1 Unsigned Integer (Hexadecimal)}
\usage{
format_asn1_unsigned_integer_hex(n)
}
\arguments{
\item{n}{A hexadecimal string representing an unsigned integer.}
}
\value{
A string containing the formatted hexadecimal representation.
}
\description{
Formats a hexadecimal-encoded unsigned integer by removing leading zeros and padding if necessary to avoid two's complement confusion.
}
