% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_key_pair.R
\name{format_key_pair}
\alias{format_key_pair}
\title{Format a Key Pair for the secp256k1 Curve}
\usage{
format_key_pair(private_key, public_key, compressed = TRUE)
}
\arguments{
\item{private_key}{A raw vector, hexadecimal string, or big integer (bigz) representing the private key.}

\item{public_key}{A hexadecimal string representing the public key.}

\item{compressed}{Logical, whether to return the public key in compressed format (default: \code{TRUE}).}
}
\value{
A list containing the hex-encoded private key and the formatted public key.
}
\description{
This function takes a key pair and returns the X9.62 compressed encoded public key and private key as a list, with each value hex-encoded.
}
\examples{
# Generate a private key using rbtc
private_key <- create_secp256k1_private_key()

# Generate the corresponding public key
public_key <- PrivKey2PubKey(private_key)

# Format the key pair
key_pair <- format_key_pair(private_key, public_key)
print(key_pair$private)
print(key_pair$public)

}
