% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_public_key.R
\name{format_public_key}
\alias{format_public_key}
\title{Format a Public Key for the secp256k1 Curve}
\usage{
format_public_key(public_key, compressed = TRUE)
}
\arguments{
\item{public_key}{A hexadecimal string representing the public key.}

\item{compressed}{Logical, whether to return the public key in compressed format (default: \code{TRUE}).}
}
\value{
A hexadecimal string containing the formatted public key.
}
\description{
This function formats the internal representation of a public key to a standard public key format used in the secp256k1 curve.
}
\examples{
private_key <- createPrivateKey()
public_key <- PrivKey2PubKey(private_key)
formatted_key <- format_public_key(public_key)

}
