% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secp256k1.R
\name{generate_keypair}
\alias{generate_keypair}
\title{Generate Key Pair}
\usage{
generate_keypair(seckey_r = NULL, output_format = c("hex", "base64", "raw")[1])
}
\arguments{
\item{seckey_r}{Optional. A raw vector containing the 32-byte secret key.}
}
\value{
A list with:
\itemize{
\item \code{seckey}: A raw vector containing the 32-byte secret key (only if no key was provided).
\item \code{pubkey}: A raw vector containing the 65-byte uncompressed public key.
}
}
\description{
Generates a secp256k1 key pair. If a secret key is provided, the function generates
the corresponding public key. If no secret key is provided, it generates a new key pair.
}
