% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secp256k1_wrapper.R
\name{get_curve_order}
\alias{get_curve_order}
\title{Get the Curve Order}
\usage{
get_curve_order()
}
\value{
A raw vector representing the curve order.
}
\description{
Retrieves the curve order from the secp256k1 elliptic curve.
}
