% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secp256k1_wrapper.R
\name{get_modulus}
\alias{get_modulus}
\title{Get the Curve Order}
\usage{
get_modulus()
}
\value{
A raw vector representing the curve order.
}
\description{
Retrieves the curve order from the secp256k1 elliptic curve.
}
