% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sin_from_public_key.R
\name{get_sin_from_private_key}
\alias{get_sin_from_private_key}
\title{get_sin_from_private_key}
\usage{
get_sin_from_private_key(private_key_hex, output_format = "base58")
}
\arguments{
\item{private_key_hex}{Character. The private key in hexadecimal format.}

\item{output_format}{Character. The output format for the SIN. Default is "base58".}
}
\value{
Character SIN
}
\description{
Derive public key from private key and generate SIN
}
\examples{
private_key_hex <- "1E99423A4ED27608A15A2616A2B9245621A89E834046E0BC3A707855853D7E58"
sin <- get_sin_from_private_key(private_key_hex)
}
