% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash3.R
\name{hash3}
\alias{hash3}
\title{General SHA3 Hashing Function}
\usage{
hash3(ba, hash_size)
}
\arguments{
\item{ba}{A raw vector representing the byte array to be hashed.}

\item{hash_size}{The size of the hash, either 256 or 512 bits.}
}
\value{
A raw vector representing the hash.
}
\description{
This function calculates a SHA3 hash of the specified size (256 or 512 bits).
}
\examples{
hash256 <- hash3(charToRaw("Hello, world!"), 256)
print(hash256)  # Outputs the SHA3-256 hash as a raw vector

hash512 <- hash3(charToRaw("Hello, world!"), 512)
print(hash512)  # Outputs the SHA3-512 hash as a raw vector

}
