% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{hash_string_key}
\alias{hash_string_key}
\title{Hash a string key}
\usage{
hash_string_key(key, n = 32)
}
\arguments{
\item{key}{A character string or raw vector to be hashed.}

\item{n}{An integer specifying the number of bytes to return from the hash. Must be between 1 and 64 (default is 32).}
}
\value{
A raw vector containing the first \code{n} bytes of the SHA3-512 hash.
}
\description{
This function takes a string key, hashes it using the SHA3-512 algorithm
and returns the first \code{n} bytes of the resulting hash.
}
\examples{
# hash_string_key("hello", 32)
# hash_string_key(charToRaw("example-key"), 16)

}
