% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sin_from_public_key.R
\name{hex_to_raw}
\alias{hex_to_raw}
\title{hex_to_raw}
\usage{
hex_to_raw(hex)
}
\arguments{
\item{hex}{Character. Hexadecimal string to convert to raw bytes.}
}
\value{
raw
}
\description{
Convert hexadecimal string to raw bytes
}
\examples{
hex <- "031fbbad2f558783b4490692f46fff26a10fc7c3633741cb450f104d598320a6d6"
raw_bytes <- hex_to_raw(hex)
}
