% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secp256k1_wrapper.R
\name{is_valid_private_key}
\alias{is_valid_private_key}
\title{Check if private key is valid}
\usage{
is_valid_private_key(seckey)
}
\arguments{
\item{seckey}{Private key (32 bytes)}
}
\value{
TRUE if the private key is valid, FALSE otherwise
}
\description{
Check if private key is valid
}
