% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_key_from_private.R
\name{keypair_from_private}
\alias{keypair_from_private}
\title{Generate Public Key from Private Key for the secp256k1 Curve}
\usage{
keypair_from_private(private_key)
}
\arguments{
\item{private_key}{A hexadecimal string or big integer (bigz) representing the private key.}
}
\value{
A list containing the big integer private key and the corresponding public key as a hexadecimal string.
}
\description{
This function generates a public key from a given private key using the secp256k1 curve.
}
\examples{
private_key <- create_valid_secp256k1_private_key()
key_pair <- public_key_from_private(private_key)
print(key_pair$private)
print(key_pair$public)

}
