% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{map_signed_to_unsigned}
\alias{map_signed_to_unsigned}
\title{Map signed to unsigned bytes}
\usage{
map_signed_to_unsigned(ba)
}
\arguments{
\item{v}{A numeric vector.}
}
\value{
A numeric vector with negative integers replaced.
}
\description{
This function takes a vector of integers and replaces negative integers
with their positive counterparts (by adding 256).
}
\examples{
# v <- c(-5, 0, 10, -128, 255)
# result <- map_signed_to_unsigned(v)
# print(result)

}
