% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{normalize_string}
\alias{normalize_string}
\title{Normalize a string}
\usage{
normalize_string(s)
}
\arguments{
\item{s}{A character string to be normalized.}
}
\value{
A character string that has been normalized to the NFKC form.
}
\description{
This function normalizes a string using the NFKC normalization form.
The normalized form of the string will result in consistent hashing.
}
\examples{
# normalized_string <- normalize_string("\u0041\u030apple")
# print(normalized_string)

}
