% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encodings.R
\name{pad_hex}
\alias{pad_hex}
\title{Pads a hexadecimal string with a leading zero if the length is odd}
\usage{
pad_hex(hex)
}
\arguments{
\item{hex}{A hexadecimal string}
}
\value{
A padded hexadecimal string with a leading zero if necessary
}
\description{
Pads a hexadecimal string with a leading zero if the length is odd
}
