% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkcs7.R
\name{pad_pkcs7}
\alias{pad_pkcs7}
\title{Pad data using PKCS7 padding}
\usage{
pad_pkcs7(data, block_size)
}
\arguments{
\item{data}{A raw vector containing the data to be padded.}

\item{block_size}{An integer specifying the block size in bytes. For AES, this is typically 16.}
}
\value{
A raw vector containing the padded data.
}
\description{
This function adds PKCS7 padding to the input data to make its length a multiple of the block size.
}
\examples{
pad_pkcs7(charToRaw("hi"), 16)
}
