% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encodings.R
\name{pad_to_length}
\alias{pad_to_length}
\title{Pads a string to a specified length with leading zeroes}
\usage{
pad_to_length(s, len)
}
\arguments{
\item{s}{A string}

\item{len}{The desired length of the string}
}
\value{
A string padded with leading zeroes
}
\description{
Pads a string to a specified length with leading zeroes
}
