% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkcs7.R
\name{pkcs7_decode}
\alias{pkcs7_decode}
\title{PKCS7 Decode}
\usage{
pkcs7_decode(k, m)
}
\arguments{
\item{k}{An integer representing the block size.}

\item{m}{A numeric vector representing the padded message (byte array).}
}
\value{
A numeric vector representing the original unpadded message.
}
\description{
Removes PKCS7 padding from a padded message. Throws an error if the padding is invalid.
}
