% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkcs7.R
\name{pkcs7_encode}
\alias{pkcs7_encode}
\title{PKCS7 Encode}
\usage{
pkcs7_encode(k, m)
}
\arguments{
\item{k}{An integer representing the block size.}

\item{m}{A numeric vector representing the message (byte array) to pad.}
}
\value{
A numeric vector representing the padded message.
}
\description{
Pads a message using PKCS7 padding based on the provided block size.
}
