% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_key_from_private.R
\name{pub_key_from_private}
\alias{pub_key_from_private}
\title{Generate Public Key from Private Key for the secp256k1 Curve}
\usage{
pub_key_from_private(private_key)
}
\arguments{
\item{private_key}{A hexadecimal string or big integer (bigz) representing the private key.}
}
\value{
A hexadecimal string representing the public key.
}
\description{
This function takes a private key (in hex string or big integer format) and returns the corresponding public key as a hex string.
}
\examples{
private_key <- create_valid_secp256k1_private_key()
public_key <- pub_key_from_private(private_key)
print(public_key)

}
