% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secp256k1.R
\name{public_key_from_private}
\alias{public_key_from_private}
\title{Derive Public Key from Private Key}
\usage{
public_key_from_private(priv_key)
}
\arguments{
\item{priv_key}{A raw vector or character string representing the private key.
This private key will be used to derive the corresponding public key.}
}
\value{
A raw vector representing the public key derived from the private key.
}
\description{
This function generates a public key from a given private key by creating a key pair.
}
\examples{
# Example usage:
# Assuming `priv_key` is your private key in raw or hex format
public_key <- public_key_from_private(priv_key)

}
