% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrypt.R
\name{random_bytes}
\alias{random_bytes}
\title{Generate Random Bytes}
\usage{
random_bytes(size)
}
\arguments{
\item{size}{Integer. The size of the random byte array to generate.}
}
\value{
A raw vector containing the random bytes.
}
\description{
Generates a random byte array of the specified size.
This will work only in environments where random byte generation is supported.
}
