% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secp256k1_wrapper.R
\name{rawToBigInteger}
\alias{rawToBigInteger}
\title{Convert Raw to Big Integer}
\usage{
rawToBigInteger(raw_data)
}
\arguments{
\item{raw_data}{A raw vector to be converted.}
}
\value{
A big integer representation of the raw vector.
}
\description{
Convert a raw vector to a big integer.
}
