% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrypt.R
\name{scrypt_encrypt}
\alias{scrypt_encrypt}
\title{Encrypt Using scrypt}
\usage{
scrypt_encrypt(
  msg,
  salt = random_bytes(16),
  n = 32768,
  r = 8,
  p = 1,
  dk_len = 32
)
}
\arguments{
\item{msg}{A raw vector or character string containing the message to encrypt.}

\item{salt}{A raw vector containing the salt to use (default random 16 bytes).}

\item{n}{Integer. CPU/memory cost factor (default 32768).}

\item{r}{Integer. Block size factor (default 8).}

\item{p}{Integer. Parallelization factor (default 1).}

\item{dk_len}{Integer. Length of the derived key (default 32).}
}
\value{
A raw vector containing the encrypted message.
}
\description{
Encrypts a message (raw bytes) using a specified salt and scrypt parameters.
Returns the encrypted message in bytes directly. The encrypted output can be
verified using the same salt and scrypt parameters.
}
