% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sha2.R
\name{sha2_256}
\alias{sha2_256}
\title{SHA-256 Hashing Function}
\usage{
sha2_256(x, output_format = c("hex", "base64", "raw")[1], input_format = NULL)
}
\arguments{
\item{x}{The input to be hashed, either a character string or a raw vector.}

\item{output_format}{The format of the output hash. Options are "hex" (default) or "base64".}

\item{input_format}{The format of the input. Options are "string" or "bytes".}
}
\value{
A character string representing the SHA-256 hash in the specified format.
}
\description{
This function calculates the SHA-256 hash of the input.
}
\examples{
sha2_256("hello")
sha2_256(charToRaw("hello"), output_format = "base64")
# => (println (alphabase/bytes->hex (sha2-256 (.getBytes "hello"))))
# 2cf24dba5fb0a30e26e83b2ac5b9e29e1b161e5c1fa7425e73043362938b9824

}
