% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sha2.R
\name{sha2_256_normalize}
\alias{sha2_256_normalize}
\title{SHA-256 Hashing Function with Normalization}
\usage{
sha2_256_normalize(s, output_format = "hex")
}
\arguments{
\item{s}{The string to be hashed.}

\item{output_format}{The format of the output hash. Options are "hex" (default) or "base64".}
}
\value{
A character string representing the SHA-256 hash in the specified format.
}
\description{
This function normalizes a string and then computes its SHA-256 hash.
}
\examples{
hash_hex <- sha2_256_normalize("Café")
print(hash_hex)  # Outputs the SHA-256 hash of the normalized string in hexadecimal format

hash_base64 <- sha2_256_normalize("Café", output_format = "base64")
print(hash_base64)  # Outputs the SHA-256 hash of the normalized string in base64 format

}
