% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sha2.R
\name{sha2_512}
\alias{sha2_512}
\title{SHA-512 Hashing Function}
\usage{
sha2_512(x, output_format = c("hex", "base64", "raw")[1], input_format = NULL)
}
\arguments{
\item{x}{The input to be hashed, either a character string or a raw vector.}

\item{output_format}{The format of the output hash. Options are "hex" (default) or "base64".}

\item{input_format}{The format of the input. Options are "string" or "bytes".}
}
\value{
A character string representing the SHA-512 hash in the specified format.
}
\description{
This function calculates the SHA-512 hash of the input.
}
\examples{
sha2_512("hello")
sha2_512(charToRaw("Hello, world!"), output_format = "base64")
## => (println (alphabase/bytes->hex (sha2-512 (.getBytes "hello"))))
## 9b71d224bd62f3785d96d46ad3ea3d73319bfbc2890caadae2dff72519673ca72323c3d99ba5c11d7c7acc6e14b8c5da0c4663475c2e5c3adef46f73bcdec043
## crypto.sha2_512("hi");
## returns: 150a14ed5bea6cc731cf86c41566ac427a8db48ef1b9fd626664b3bfbb99071fa4c922f33dde38719b8c8354e2b7ab9d77e0e67fc12843920a712e73d558e197.
}
