% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jws.R
\name{sign}
\alias{sign}
\title{Sign a message using secp256k1 private key}
\usage{
sign(signing_input, signing_key)
}
\arguments{
\item{signing_input}{The message to be signed as a raw vector or string.}

\item{signing_key}{The secp256k1 private key used for signing.}
}
\value{
A signature as a raw vector.
}
\description{
This function signs the provided input using a secp256k1 private key and SHA-256 hash algorithm.
}
\examples{
sign("message to sign", my_signing_key)
}
