% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secp256k1_wrapper.R
\name{sign_hash}
\alias{sign_hash}
\title{Sign a message hash}
\usage{
sign_hash(seckey, hash)
}
\arguments{
\item{seckey}{Private key (32 bytes)}

\item{hash}{Message hash (32 bytes)}
}
\value{
A raw vector of the signature (64 bytes)
}
\description{
Sign a message hash
}
