% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secp256k1.R
\name{sign_message}
\alias{sign_message}
\title{Sign a message hash}
\usage{
sign_message(msg, priv_key, output_format = c("hex", "base64", "raw")[1])
}
\arguments{
\item{msg}{A raw vector containing the 32-byte message hash or the original message as a character string.}

\item{priv_key}{A raw vector containing the 32-byte private key or the private key as a hexadecimal string.}
}
\value{
The signature as a hexadecimal string.
}
\description{
This wrapper function calls on the corresponding C function to sign a 32-byte
message hash using a provided private key. If the provided message is a
character string the sha2_256() hash of the message is used for signing.
The resulting signature is in DER encoded format prepended by a recovery byte.
}
\examples{
# sig <- sign_message("hi", "6a5f415f49986006815ae7887016275aac8ffb239f9a2fa7172300578582b6c2")

}
