% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbtc_approach.R
\name{sign_message_rbtc}
\alias{sign_message_rbtc}
\title{Sign a Message with a Private Key using ECDSA and secp256k1 (rbtc)}
\usage{
sign_message_rbtc(message, private_key)
}
\arguments{
\item{message}{A string or raw byte array representing the message to be signed.}

\item{private_key}{A hexadecimal string representing the private key.}
}
\value{
A hexadecimal string representing the ECDSA signature.
}
\description{
This function signs a message with a provided private key using the secp256k1 curve.
The message can be a string or raw byte array. The private key must be provided as a hexadecimal string.
}
\examples{
private_key <- generate_secp256k1_private_key()
message <- "Hello, blockchain!"
signature <- sign_message_rbtc(message, private_key)
print(signature)

}
