% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign.R
\name{sign_message_sodium}
\alias{sign_message_sodium}
\title{Sign a Message with a Private Key using ECDSA and secp256k1 (Sodium)}
\usage{
sign_message_sodium(message, private_key)
}
\arguments{
\item{message}{A string or raw byte array representing the message to be signed.}

\item{private_key}{A hexadecimal string or raw byte array representing the private key.}
}
\value{
A raw byte array representing the ECDSA signature.
}
\description{
This function signs a message with a provided private key using the secp256k1 curve.
The message can be a string or raw byte array. The private key must be provided as a hexadecimal string or raw byte array.
}
\examples{
private_key <- "1E99423A4ED27608A15A2616A2B9245621A89E834046E0BC3A707855853D7E58"
message <- "Hello, blockchain!"
signature <- sign_message_sodium(message, private_key)
print(sodium::bin2hex(signature))

}
