% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{string_to_byte_array}
\alias{string_to_byte_array}
\title{Convert a string to a byte array}
\usage{
string_to_byte_array(s)
}
\arguments{
\item{s}{A character string or a raw vector. The string will be normalized before being converted to a byte array.}
}
\value{
A raw vector representing the byte array.
}
\description{
This function normalizes a string and then converts it to a byte array.
If the input is already a byte array, it returns the original value.
}
\examples{
# byte_array <- string_to_byte_array("Åpple")
# print(byte_array)

# raw_input <- as.raw(1:5)
# byte_array <- string_to_byte_array(raw_input)
# print(byte_array)

}
