% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkcs7.R
\name{unpad_pkcs7}
\alias{unpad_pkcs7}
\title{Remove PKCS7 padding from data}
\usage{
unpad_pkcs7(data, block_size)
}
\arguments{
\item{data}{A raw vector containing the padded data.}

\item{block_size}{An integer specifying the block size in bytes. For AES, this is typically 16.}
}
\value{
A raw vector containing the data with padding removed.
}
\description{
This function removes PKCS7 padding from the input data. It verifies that the padding is correct.
}
\examples{
unpad_pkcs7(pad_pkcs7(charToRaw("hi"), 16), 16)
}
