% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secp256k1.R
\name{valid_private}
\alias{valid_private}
\title{Validate private key}
\usage{
valid_private(private_key_hex)
}
\arguments{
\item{private_key_hex}{A string representing the private key in hexadecimal format.}
}
\value{
An integer (1 if valid, 0 otherwise).
}
\description{
This wrapper function calls its corresponding C function which
checks if a provided private key is valid on the secp256k1 curve.
}
\examples{
# if (valid_private("6a5f415f49986006815ae7887016275aac8ffb239f9a2fa7172300578582b6c2")) {
#   print("Private key is valid")
# }

}
