% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_private_key.R
\name{valid_private_key}
\alias{valid_private_key}
\title{Validate a Private Key for the secp256k1 Curve}
\usage{
valid_private_key(private_key)
}
\arguments{
\item{private_key}{A \code{bigz} object representing the private key to be validated.}
}
\value{
A logical value (\code{TRUE} or \code{FALSE}) indicating whether the private key is valid.
}
\description{
This function checks whether a given private key is valid for the secp256k1 curve. A valid private key must be a big integer that falls within the range \link{1, modulus - 1}, where modulus is the order of the secp256k1 curve.
}
\examples{
# Example private key in hexadecimal
private_key <- "0521b34877d71a6c7bba4cc53c5ffdbf5023c3b1fb71f7b1a384330fe6e9986d"

# Convert to bigz and validate
private_bn <- gmp::as.bigz(paste0("0x", private_key))
is_valid <- valid_private_key(private_bn)
print(is_valid)  # This should print TRUE if the key is valid

}
