% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secp256k1.R
\name{verify_signature}
\alias{verify_signature}
\title{Verify a Signature from a Hash}
\usage{
verify_signature(pub_key, message, sig)
}
\arguments{
\item{pub_key}{A character string representing the public key in hexadecimal format.}

\item{message}{A character string representing the original message.}

\item{signature}{A character string representing the signature in hexadecimal format,
assumed to be DER-encoded with a recovery byte.}
}
\value{
TRUE if the signature is valid for the given public key and hash,
otherwise throws an error.
}
\description{
Verifies that a signature is valid for a given public key and hash.
}
\examples{
# Example usage:
# verify(pub_key, message, sig)

}
