% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accountID.R
\name{account_id_from_private}
\alias{account_id_from_private}
\title{Generate Account ID (SIN) from Private Key}
\usage{
account_id_from_private(private_key)
}
\arguments{
\item{private_key}{A hexadecimal string or big integer (bigz) representing the private key.}
}
\value{
A string representing the account ID (SIN).
}
\description{
This function takes a private key as input and derives the corresponding account ID (SIN).
}
\examples{
private_key <- create_valid_secp256k1_private_key()
account_id <- account_id_from_private(private_key)
print(account_id)

}
