% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accountID.R
\name{account_id_from_public}
\alias{account_id_from_public}
\title{Generate Account ID (SIN) from Public Key}
\usage{
account_id_from_public(public_key)
}
\arguments{
\item{public_key}{A hexadecimal string representing the public key.}
}
\value{
A string representing the account ID (SIN).
}
\description{
This function takes a public key as input and derives the corresponding account ID (SIN).
}
\examples{
public_key <- pub_key_from_private(create_valid_secp256k1_private_key())
account_id <- account_id_from_public(public_key)
print(account_id)

}
