% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secp256k1.R
\name{account_id_from_public}
\alias{account_id_from_public}
\title{Generate Account ID from Public Key}
\usage{
account_id_from_public(pub_key)
}
\arguments{
\item{pub_key}{A raw vector or character string representing the public key.
This public key will be used to derive the account ID.}
}
\value{
A character string representing the account ID in the specified format
(typically base58).
}
\description{
This function generates an account identifier (SIN) from a given public key.
}
\examples{
# Example usage:
# Assuming `pub_key` is your public key in raw or hex format
account_id <- account_id_from_public(pub_key)

}
