% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes.R
\name{aes_encrypt}
\alias{aes_encrypt}
\title{AES Encryption Wrapper Function}
\usage{
aes_encrypt(x, iv, key, output_format = "hex")
}
\arguments{
\item{x}{The input to be encrypted, either a character string or a raw vector.}

\item{iv}{A raw vector representing the initialization vector (IV).}

\item{key}{The encryption key as a character string or raw vector. It will be hashed to 256 bits if provided as a string.}

\item{output_format}{The format of the output. Options are "hex" (default), "base64", or "none" for raw bytes.}
}
\value{
The encrypted data in the specified format.
}
\description{
Encrypts the input using AES encryption in CBC mode with PKCS7 padding, with default output format as hexadecimal.
}
\examples{
aes_encrypt("Encrypt this message!", charToRaw("thisisaniv123456"), "mysecretpassword")

}
