% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encodings.R
\name{biginteger_to_bytes}
\alias{biginteger_to_bytes}
\title{Convert BigInteger to Byte Array}
\usage{
biginteger_to_bytes(bn)
}
\arguments{
\item{bn}{A GMP big integer (from the \code{gmp} library in R).}
}
\value{
A raw byte array representing the big integer.
}
\description{
This function converts a GMP big integer to a raw byte array.
}
