% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{coerce_input_format}
\alias{coerce_input_format}
\title{Coerce input format}
\usage{
coerce_input_format(x)
}
\arguments{
\item{x}{The input to be checked, which can be a character string or raw vector.}
}
\value{
A character string indicating the format, either "string" or "bytes".
}
\description{
This function checks whether the input is a string or raw bytes and returns
a label indicating the format.
}
\examples{
# input_format <- coerce_input_format("Hello, world!")
# print(input_format)

# input_format <- coerce_input_format(as.raw(1:5))
# print(input_format)

}
