% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes.R
\name{decrypt_aes_cbc}
\alias{decrypt_aes_cbc}
\title{Decrypt data using AES}
\usage{
decrypt_aes_cbc(iv, key, encrypted_data)
}
\arguments{
\item{iv}{A raw vector representing the initialization vector.}

\item{key}{A raw vector representing the key for AES decryption. It should be exactly 16, 24, or 32 bytes.}

\item{encrypted_data}{A raw vector representing the data to be decrypted.}
}
\value{
The character string representation of the decrypted data.
}
\description{
This function decrypts a message using AES decryption in CBC mode with PKCS7 padding.
It receives the necessary input from the "decrypt" function after all the
necessary type-checking and conversions have been done.
}
