% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jws.R
\name{deserialize_jws}
\alias{deserialize_jws}
\title{Deserialize a JWS Compact Serialization}
\usage{
deserialize_jws(jws)
}
\arguments{
\item{jws}{A JWS compact serialization string.}
}
\value{
A list containing the decoded header, payload, and signature.
}
\description{
This function splits a JWS Compact Serialization into its component parts:
header, payload, and signature.
It decodes the Base64 URL encoded parts back into readable formats.
}
\examples{
deserialize_jws(my_jws_string)

}
