% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secp256k1_wrapper.R
\name{deterministic_generate_k}
\alias{deterministic_generate_k}
\title{Deterministic Generation of k}
\usage{
deterministic_generate_k(hash_ba, priv_key)
}
\arguments{
\item{hash_ba}{A raw vector representing the hash (should be 32 bytes).}

\item{priv_key}{A raw vector representing the private key.}
}
\value{
A big integer representing the generated k value.
}
\description{
Deterministically generate a random number in accordance with RFC 6979.
Provided hash should have 256 bits to align with the secp256k1 curve.
}
