% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes.R
\name{encrypt}
\alias{encrypt}
\title{Encrypt Data with AES/CBC/PKCS7 Padding}
\usage{
encrypt(
  x,
  key,
  iv = c(6, 224, 71, 170, 241, 204, 115, 21, 30, 8, 46, 223, 106, 207, 55, 42),
  output_format = "hex"
)
}
\arguments{
\item{x}{The input data to encrypt. This can be a character string or a raw vector.}

\item{key}{The encryption key. This can be a character string (hashed into a 256-bit key) or a raw vector.}

\item{iv}{An optional vector of unsigned bytes of size 16 to be used as the initialization vector. Defaults to a predefined IV.}

\item{output_format}{The desired format for the encrypted output: "hex", "base64", or "none".}
}
\value{
The encrypted data in the specified output format.
}
\description{
Encrypt Data with AES/CBC/PKCS7 Padding
}
