% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes.R
\name{encrypt_aes_cbc}
\alias{encrypt_aes_cbc}
\title{Encrypt data using AES}
\usage{
encrypt_aes_cbc(iv, key, data)
}
\arguments{
\item{iv}{A raw vector of length 16 representing the initialization vector.}

\item{key}{A raw vector of length 32 representing the AES key.}

\item{data}{A raw vector representing the plaintext data/message to encrypt.}
}
\value{
A raw vector representing the encrypted data.
}
\description{
This function performs AES encryption in CBC mode with PKCS#7 padding.
It is called from the "encrypt" function below after all the necessary
conversions and type-checking has been done.
}
