% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secp256k1.R
\name{generate_seckey}
\alias{generate_seckey}
\title{Generate Secret Key}
\usage{
generate_seckey(output_format = c("hex", "base64", "raw")[1])
}
\value{
A raw vector containing the 32-byte secret key.
}
\description{
Generates a random 32-byte secret key that is valid on the secp256k1 curve.
}
