% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sin_from_public_key.R
\name{get_sin_from_public_key}
\alias{get_sin_from_public_key}
\title{get_sin_from_public_key}
\usage{
get_sin_from_public_key(pub_key_hex, output_format = "base58")
}
\arguments{
\item{pub_key_hex}{Character. The public key in hexadecimal format.}

\item{output_format}{Character. The output format for the SIN. Default is "base58".}
}
\value{
character
}
\description{
Generate SIN from public key
}
\examples{
pub_key_hex <- "031fbbad2f558783b4490692f46fff26a10fc7c3633741cb450f104d598320a6d6"
sin <- get_sin_from_public_key(pub_key_hex)
}
