% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secp256k1.R
\name{get_sin_from_public_key}
\alias{get_sin_from_public_key}
\title{Generate a SIN (Secure Identity Number) from a public key}
\usage{
get_sin_from_public_key(pub_key, output_format = "base58")
}
\arguments{
\item{pub_key}{A character string (hexadecimal format) or raw vector representing the public key.}

\item{output_format}{A character string specifying the output format.
Can be "hex" (hexadecimal string), "raw" (raw byte vector), or "base58" (Base58Check encoded string).
Default is "base58".}
}
\value{
The SIN in the specified format. The output will be:
\itemize{
\item a hexadecimal string if \code{output_format} is "hex",
\item a raw vector if \code{output_format} is "raw",
\item a Base58Check encoded string if \code{output_format} is "base58".
}
}
\description{
This function generates a SIN from a given public key by leveraging the
SHA-256 and RIPEMD-160 hash functions.
}
\examples{
# pub_key_hex <- "02991719b37817f6108fc8b0e824d3a9daa3d39bc97ecfd4f8bc7ef3b71d4c6391"
# sin <- get_sin_from_public_key(pub_key_hex, output_format = "base58")

}
