% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accountID.R
\name{get_sin_from_public_key}
\alias{get_sin_from_public_key}
\title{Derive Single Identity Number (SIN) from Public Key}
\usage{
get_sin_from_public_key(public_key)
}
\arguments{
\item{public_key}{A hexadecimal string representing the public key.}
}
\value{
A string representing the Single Identity Number (SIN).
}
\description{
This function takes a public key as input, hashes it, and encodes the result to derive the Single Identity Number (SIN).
}
\examples{
public_key <- pub_key_from_private(create_valid_secp256k1_private_key())
sin <- get_sin_from_public_key(public_key)
print(sin)

}
