% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{normalize_string}
\alias{normalize_string}
\title{Normalize a String for Consistent Hashing}
\usage{
normalize_string(s)
}
\arguments{
\item{s}{A character string to be normalized.}
}
\value{
A character string that has been normalized to the NFKC form.
}
\description{
This function normalizes a string using the NFKC normalization form.
}
\examples{
normalized_string <- normalize_string("Café")
print(normalized_string)

}
