% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secp256k1.R
\name{public_key_from_message}
\alias{public_key_from_message}
\title{Recover Public Key from Signature}
\usage{
public_key_from_message(hex_signature, msg)
}
\arguments{
\item{hex_signature}{A string representing the ECDSA signature in hexadecimal format, including the recovery byte.}

\item{msg}{A character string or raw vector representing the message hash that was signed.}
}
\value{
A raw vector containing the 65-byte compressed public key if the recovery is successful.
}
\description{
Recovers the public key associated with a given ECDSA signature and message hash on the secp256k1 curve.
}
