% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secp256k1.R
\name{public_key_from_private}
\alias{public_key_from_private}
\title{Derive a public key from a private key}
\usage{
public_key_from_private(priv_key)
}
\arguments{
\item{priv_key}{A raw vector or hexadecimal character string representing the private key.}
}
\value{
A raw vector representing the public key derived from the private key.
}
\description{
This function generates a public key from a given private key by creating a key pair.
}
\examples{
# priv_key = "6a5f415f49986006815ae7887016275aac8ffb239f9a2fa7172300578582b6c2"
# public_key <- public_key_from_private(priv_key)

}
