% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripemd.R
\name{ripemd_160}
\alias{ripemd_160}
\title{RIPEMD-160 Hashing Function}
\usage{
ripemd_160(ba, output_format = "hex")
}
\arguments{
\item{ba}{A raw vector representing the byte array to be hashed.}
}
\value{
A raw vector representing the RIPEMD-160 hash.
}
\description{
This function calculates the RIPEMD-160 hash of the input byte array.
}
\examples{
hash_ripemd160 <- ripemd_160(charToRaw("hi there!"))

}
