% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashFunctions.R
\name{sha2_512}
\alias{sha2_512}
\title{SHA-512 Hashing Function}
\usage{
sha2_512(x, output_format = "hex", input_format = NULL)
}
\arguments{
\item{x}{The input to be hashed, either a character string or a raw vector.}

\item{output_format}{The format of the output hash. Options are "hex" (default) or "base64".}

\item{input_format}{The format of the input. Options are "string" or "bytes".}
}
\value{
A character string representing the SHA-512 hash in the specified format.
}
\description{
This function calculates the SHA-512 hash of the input.
}
\examples{
sha2_512("Hello, world!")
sha2_512(charToRaw("Hello, world!"), output_format = "base64")

}
