% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashFunctions.R
\name{sha3_256}
\alias{sha3_256}
\title{SHA3-256 Hashing Function}
\usage{
sha3_256(x, output_format = "hex", input_format = NULL)
}
\arguments{
\item{x}{The input to be hashed, either a character string or a raw vector.}

\item{output_format}{The format of the output hash. Options are "hex" (default) or "base64".}

\item{input_format}{The format of the input. Options are "string" or "bytes".}
}
\value{
A character string representing the SHA3-256 hash in the specified format.
}
\description{
This function calculates the SHA3-256 hash of the input.
}
\examples{
sha3_256("Hello, world!")
sha3_256(charToRaw("Hello, world!"), output_format = "base64")

}
