% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash3.R
\name{sha3_512}
\alias{sha3_512}
\title{SHA3-512 Hashing Function}
\usage{
sha3_512(x, output_format = "hex", input_format = NULL)

sha3_512(x, output_format = "hex", input_format = NULL)
}
\arguments{
\item{x}{The input to be hashed, either a character string or a raw vector.}

\item{output_format}{The format of the output hash. Options are "hex" (default) or "base64".}

\item{input_format}{The format of the input. Options are "string" or "bytes".}

\item{ba}{A raw vector representing the byte array to be hashed.}
}
\value{
A raw vector representing the SHA3-512 hash.

A character string representing the SHA3-512 hash in the specified format.
}
\description{
This function calculates the SHA3-512 hash of the input byte array.

This function calculates the SHA3-512 hash of the input and returns it in the specified output format.
}
\examples{
hash512 <- sha3_512(charToRaw("Hello, world!"))
print(hash512)  # Outputs the SHA3-512 hash as a raw vector

sha3_512("Hello, world!")
sha3_512(charToRaw("Hello, world!"), output_format = "base64")

}
