% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sha3.R
\name{sha3_512}
\alias{sha3_512}
\title{SHA3-512 Hashing Function with Output Format}
\usage{
sha3_512(x, output_format = c("hex", "base64", "raw"), input_format = NULL)
}
\arguments{
\item{x}{The input to be hashed, either a character string or a raw vector.}

\item{output_format}{The format of the output hash. Options are "hex" (default) or "base64".}

\item{input_format}{The format of the input. Options are "string" or "bytes".}
}
\value{
A character string representing the SHA3-512 hash in the specified format.
}
\description{
This function calculates the SHA3-512 hash of the input and returns it in the specified output format.
}
\examples{
sha3_512("hello")
sha3_512(charToRaw("hello"), output_format = "base64")
# fluree.crypto.sha3=> (println (alphabase/bytes->hex (sha3-512 (.getBytes "hello"))))
# 75d527c368f2efe848ecf6b073a36767800805e9eef2b1857d5f984f036eb6df891d75f72d9b154518c1cd58835286d1da9a38deba3de98b5a53e5ed78a84976
}
