% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secp256k1_wrapper.R
\name{sign_hash}
\alias{sign_hash}
\title{Sign a Hash}
\usage{
sign_hash(hash_ba, private_bn, recovery_byte = TRUE)
}
\arguments{
\item{hash_ba}{A raw vector representing the hash to be signed.}

\item{private_bn}{A numeric representing the private key.}

\item{recovery_byte}{Logical indicating whether to compute the recovery byte.}
}
\value{
A character string representing the hexadecimal encoded DER signature.
}
\description{
Returns a signature for the given hash using the provided private key.
Utilizes deterministic generation of k according to RFC 6979.
}
