% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secp256k1.R
\name{valid_private}
\alias{valid_private}
\title{Validate Private Key}
\usage{
valid_private(private_key_hex)
}
\arguments{
\item{private_key_hex}{A string representing the private key in hexadecimal format.}
}
\value{
An integer scalar (1 if valid, 0 otherwise).
}
\description{
Checks if a provided private key (hexadecimal string) is valid on the secp256k1 curve.
}
