% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jws.R
\name{verify_jws}
\alias{verify_jws}
\title{Verify a JWS signature}
\usage{
verify_jws(jws, public_key)
}
\arguments{
\item{jws}{A JWS compact serialization string to be verified.}

\item{public_key}{The secp256k1 public key used for verifying the signature
.}
}
\value{
A list containing the payload and public key if verification is successful, otherwise an error is raised.
}
\description{
This function verifies the signature of a JWS using the secp256k1 public key.
It decodes the JWS, reconstructs the signing input, and checks if the
signature matches the input.
}
\examples{
verify(my_jws_string, my_public_key)
}
